float int

+++ docker/sqrt.py
@@ -3,7 +3,7 @@ import sys
 import math
 f = open(sys.argv[1].strip())
 x = int(f.read())
-y = math.sqrt(x)
+y = int(math.sqrt(x))
 
 
 f = open(sys.argv[2].strip(),"w")
